Move Azure from com.microsoft.azure to com.microsoft.azure.management
