+++ azure-samples/src/main/java/com/microsoft/azure/management/batch/samples/ManageBatchAccount.java
@@ -7,7 +7,7 @@
 
 package com.microsoft.azure.management.batch.samples;
 
-import com.microsoft.azure.Azure;
+import com.microsoft.azure.management.Azure;
 import com.microsoft.azure.management.batch.AccountKeyType;
 import com.microsoft.azure.management.batch.Application;
 import com.microsoft.azure.management.batch.ApplicationPackage;
