+++ azure-samples/src/main/java/com/microsoft/azure/management/keyvault/samples/ManageKeyVault.java
@@ -7,7 +7,7 @@
 
 package com.microsoft.azure.management.keyvault.samples;
 
-import com.microsoft.azure.Azure;
+import com.microsoft.azure.management.Azure;
 import com.microsoft.azure.credentials.ApplicationTokenCredentials;
 import com.microsoft.azure.management.keyvault.KeyPermissions;
 import com.microsoft.azure.management.keyvault.SecretPermissions;
