+++ azure-samples/src/main/java/com/microsoft/azure/management/network/samples/ManageInternalLoadBalancer.java
@@ -7,7 +7,7 @@
 
 package com.microsoft.azure.management.network.samples;
 
-import com.microsoft.azure.Azure;
+import com.microsoft.azure.management.Azure;
 import com.microsoft.azure.management.compute.AvailabilitySet;
 import com.microsoft.azure.management.compute.KnownLinuxVirtualMachineImage;
 import com.microsoft.azure.management.compute.VirtualMachine;
