+++ azure-samples/src/main/java/com/microsoft/azure/management/resources/samples/ManageResource.java
@@ -9,7 +9,7 @@ package com.microsoft.azure.management.resources.samples;
 
 import java.io.File;
 
-import com.microsoft.azure.Azure;
+import com.microsoft.azure.management.Azure;
 import com.microsoft.azure.management.resources.fluentcore.arm.Region;
 import com.microsoft.azure.management.resources.fluentcore.utils.ResourceNamer;
 import com.microsoft.azure.management.storage.SkuName;
