+++ azure-samples/src/main/java/com/microsoft/azure/management/storage/samples/ManageStorageAccount.java
@@ -7,7 +7,7 @@
 
 package com.microsoft.azure.management.storage.samples;
 
-import com.microsoft.azure.Azure;
+import com.microsoft.azure.management.Azure;
 import com.microsoft.azure.management.resources.fluentcore.arm.Region;
 import com.microsoft.azure.management.samples.Utils;
 import com.microsoft.azure.management.storage.StorageAccount;
