+++ azure/src/test/java/com/microsoft/azure/management/TestNetwork.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for
  * license information.
  */
-package com.microsoft.azure;
+package com.microsoft.azure.management;
 
 import java.util.List;
