+++ azure/src/test/java/com/microsoft/azure/management/TestVirtualMachineSsh.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.azure;
+package com.microsoft.azure.management;
 
 import com.jcraft.jsch.JSch;
 import com.jcraft.jsch.Session;
